requirements_Version5.txt
torch>=2.2
numpy>=1.24
Pillow>=9.5
tqdm>=4.65
onnx>=1.14
torchvision>=0.17
